+++ cdm/cpp/properties/SEFunctionElectricPotentialVsTime.h
@@ -29,14 +29,14 @@ protected:
   static void Serialize(const SEFunctionElectricPotentialVsTime& src, cdm::FunctionElectricPotentialVsTimeData& dst);
 public:
 
-  double                               GetIndependentValue(unsigned int index) = delete;
-  virtual double                       GetTimeValue(unsigned int index, const TimeUnit& unit);
+  double                               GetIndependentValue(size_t index) = delete;
+  virtual double                       GetTimeValue(size_t index, const TimeUnit& unit);
   virtual std::vector<double>&         GetTime();
   virtual const TimeUnit*              GetTimeUnit();
   virtual void                         SetTimeUnit(const TimeUnit& unit);
 
-  double                               GetDependentValue(unsigned int index) = delete;
-  virtual double                       GetElectricPotentialValue(unsigned int index, const ElectricPotentialUnit& unit);
+  double                               GetDependentValue(size_t index) = delete;
+  virtual double                       GetElectricPotentialValue(size_t index, const ElectricPotentialUnit& unit);
   virtual std::vector<double>&         GetElectricPotential();
   virtual const ElectricPotentialUnit* GetElectricPotentialUnit();
   virtual void                         SetElectricPotentialUnit(const ElectricPotentialUnit& unit);
