+++ cdm/cpp/properties/SEFunctionVolumeVsTime.h
@@ -29,14 +29,14 @@ protected:
   static void Serialize(const SEFunctionVolumeVsTime& src, cdm::FunctionVolumeVsTimeData& dst);
 public:
 
-  double                       GetIndependentValue(unsigned int index) = delete;
-  virtual double               GetTimeValue(unsigned int index, const TimeUnit& unit);
+  double                       GetIndependentValue(size_t index) = delete;
+  virtual double               GetTimeValue(size_t index, const TimeUnit& unit);
   virtual std::vector<double>& GetTime();
   virtual const TimeUnit*      GetTimeUnit();
   virtual void                 SetTimeUnit(const TimeUnit& unit);
 
-  double                       GetDependentValue(unsigned int index) = delete;
-  virtual double               GetVolumeValue(unsigned int index, const VolumeUnit& unit);
+  double                       GetDependentValue(size_t index) = delete;
+  virtual double               GetVolumeValue(size_t index, const VolumeUnit& unit);
   virtual std::vector<double>& GetVolume();
   virtual const VolumeUnit*    GetVolumeUnit();
   virtual void                 SetVolumeUnit(const VolumeUnit& unit);
